Server is still a WIP but client seems to be done for now. Game, events, and verbs added, and parse event added to events.
Having trouble deciding where players should be stored. Should they just be a dictionary stored on the server, or should they be stored in the world dictionary as well?

Actually, I know. They should be in the world dict. It is kept in RAM while the game is running and their messages can, when online, be stored with them. Their locations will be kept up to date in themselves, and will also be stored in the room in which they are present.

When offline, the player will be labeled as such and moved to an offline players area.

I still cannot figure out how actors should be handled other than players. Like skeletons that move around. How should mobs work?